libgit: don't lock during the whole reset, just make a file
Resets can take way longer than the server lock expiration time.  So
just lock while creating a "working" file, and set the mtime on that
file to our best estimate of a common time derived based on our time
offset with the mdserver we're connected to.

Then resets can take up to an hour, and when they're done the worker
can delete the file and write any errors into a new "lasterr" file.