libfuse: avoid doing full TLF loads when `.localized` is requested
In particular, specify a set of directory entry names that should NOT
cause a TLF to be fully loaded and identified on a lookup.  If the
directory is not yet loaded and one of these names are looked-up, then
ENOENT will be returned automatically.  This is to avoid unnecessary
loading and tracker popups when listing the folder list directories.
For example, when macOS finder opens `/keybase/private`, it looks up
`.localized` in every TLF subdirectory to see if it should translate
the TLF folder name or not, which can cause a tracker popup storm

Issue: KBFS-2649