redirector: add a /keybase entry when KBFS is not running
Having an entry here will prevent other mounts from mounting over it,
since it will be non-empty.  That means future `run_keybase` calls
won't keep spinning up new redirector processes, due to the fact that
the running-as-root mount process sees an empty mountpoint.

For now, I just did the simplest thing: a symlink called
"KBFS_NOT_RUNNING" that points to /dev/null.  In the future, it might
be nice to have it point to a README file somewhere on disk, but that
would be a bigger undertaking.

Issue: #1510