keybase_service_base: call new service RPC to get ALL revoked keys
We previously had no way of getting all the pre-reset keys of a user,
which led to situations where team folders are unreadable when they
are last written by a user who then reset their account.

The service started exposing a V2 RPC that gives us all past
incarnations of the user, so we can check all their keys, using the
"reset" time as the revocation time.

Issue: KBFS-2842