libkbfs: stop QR early when an unverifiable TLF
Apparently some users have sigchains with revocations that are missing
the root sequence number of when the revocation happened.  In this
case, the service tells us they were revoked at seqno 0.  The API
server rejects this as an invalid seqno, and gives back an opaque
error type, which the mdserver sends back to the client.  A background
process like QR treats this as a transient error and tries the TLF
again, over and over.

This issue will go away once the service lands CORE-7806, but for now
we can prevent this from happening forever by catching the problem
early and returning a specific error type that will stop QR early.