keybase_service_base: use `FindNextMerkleRootAfterRevoke`
Instead of relying on the `PrevMerkleRootSigned` field in the
revocation (which is the seqno that the client performing the
revocation saw before pushing it), use the service to look up the
exact merkle root that first contains the revocation.

This is more accurate and gets around the issue where some old
revocations have a 0 root seqno.