md_ops: iterate correctly when decrypting merkle leaf
Forgot to set the next MD when iterating in a loop to try to decrypt a
Merkle leaf.  This caused infinite loops whenever we were trying to
verify an MD written by a revoked key, and the leaf we were using to
verify it was encrypted for a key generation that was more than 10 MD
updates away from the MD being verified.

Issue: KBFS-3019