tlf_journal: return an estimated end time for flushing bytes
This adds an exponentially-weighted moving average for estimated
upload speed in bytes per second to each TLF journal, and adds an end
time prediction to the journal status based on that estimate.  For the
overall journal status, it uses the latest end-time estimate among all
the TLFs.

If the status is obtained before any uploads for that TLF have
completed, it predicts an arbitrary end estimate of 10 seconds from
now.

If multiple TLFs are uploading at the same time, the hope is that
their bandwidth estimates will decrease because of the increased
competition for the TCP channel among the TLF journals.  If this
doesn't work well, we change this to have a system-wide bandwidth
estimate, but since all the TLFs upload data in parallel, it's not
clear exactly how this should work.

Note that a traditional exponentially-weighted moving average would
probably input sample measurements at constant time intervals, but
since uploading is sporadic we can't really do that.  Instead, our
decay is exponential in the number of samples we've taken, so it could
have a long memory in terms of time, if there haven't been many
uploads recently.

Issue: KBFS-3044