rekey_queue: put queue size into mode, and drop reqs when full
Launching a goroutine for rekey requests when the queue is full
doesn't make a lot of sense if we're trying to conserve memory (like
on a phone), since it probably uses a lot more memory than just
expanding the queue.  So this PR just drops requests when the queue is
full, and increases the queue size on desktop to handle more requests
at a time.  Any dropped requests will get processed again at the next
rekey interval.

This only changes behavior for people with more than 2K TLFs, which is
probably only two users.

Issue: KBFS-3072