keybase_service_util: don't block on initializing FBOs on login
`serviceLoggedIn` can be called as a side-effect of the mdserver
`OnConnect` call, if that is the first time the current session is
discovered.  Initializing FBOs depends on the mdserver, and so that
can lead to a deadlock.

Instead, initialize them in the background in that case, while the
other cases can still wait for the initialization to complete.

Issue: KBFS-3155