kbfs_ops: on TLF history init, don't run TLF identifies
Just start up the chat monitoring, so we don't trigger any unintended
popups.

If there are unflushed journals for some TLFs, we could theoretically
trigger popups for them since they are initialized differently, but in
that case the user would probably have recently written to the TLF and
so would have already gone through the identify flow.

Issue: CORE-8290