cr_chains: fill in paths for deleted nodes
When deleting a file AND its parent directory, the finalPath of the
corresponding ops weren't getting filled in.  This was leading to a
crash in both the journal status code to show unflushed paths, as well
as a crash in the edit history notification code, since the finalPath
is required to construct the edit message.

This commit searches for the first parent that wasn't deleted (and
hence has a path), and sets the final paths for ops of deleted
directories.

Issue: keybase/client#12817