prev_revisions: space out revisions more evenly
When playing around with time travel, I noticed that our revision
lists for files that had only a few revisions (< 100) ended up
bunching revisions together near the end of the list.  For example, if
a file had only 10 revisions, the revision count of the list is:

[1, 7, 8, 9, 10]

That is, the list is heavily weighted to older revisions, when a
proper skip list should have more newer revisions.

So this tweaks the algorithm to drop entries that are too close in
count to their right neighbor.  I picked 20% of the right-neighbor's
min-count, as it seemed to produce decent results.

I also adjusted the min counts a little bit, to keep around more newer
revisions.

With the new algorithm, here's what the counts for N revisions looks
like:

10: [1, 2, 3, 4, 10]
20: [1, 2, 3, 5, 25]
50: [1, 7, 15, 30, 50]
75: [1, 8, 16, 55, 75]

Issue: KBFS-3281