folder_branch_ops: update directory mtime/ctime on add/remove
Ever since we added in-memory operation batching in `folderBranchOps`,
and moved old simple tests to the DSL tests, we haven't been properly
updating directory mtime/ctimes when adding or removing an entry
within that directory.  We were setting the times properly in the
in-memory cache, so they appeared to be correct until the next sync
happened.  But during that sync, it turns out the
`folderUpdatePrepper` was relying on the times in the provided local
block cache object, and we weren't adding the parent directories into
that cache (unless they themselves were dirty also).

So this commit adds those parent directories to the local block cache,
and keeps the root directory entry up-to-date directly.  It also adds
some tests to catch this in the future.

Issue: KBFS-3281