file_data: generalize functions for getting blocks
When we introduce indirect directory blocks, we'll want to reuse the
logic for navigating a tree of blocks.  Do this by adding generic
offset types and some new helper functions to the Block interface, and
refactoring functions from `fileData` into a new generic `blockTree`
struct that can operate on any type of block.

This commit shouldn't change any logicc, it's just a refactor

Issue: KBFS-3299