folder_block_ops: use `dirData` and get rid of `deCache`
`deCache` has been outdated for a while now.  It used to be important
when we allowed files to be synced individually, so that one file's
directory entry changes could be preserved in memory, while flushing
only another file's entry in the same directory.  Now that we only
support `SyncAll`, it can be removed in favor of just keeping a single
view of dirty `DirBlock`s in the dirty block cache.

The transition to `dirData` seemed like a good time to do this, so I
combined the two.  Now all changes to directories are synced to the
dirty blocked cache, and all changed to the root directory entry are
kept in memory in the `folderBlockOps` instance.  This allowed a bunch
of code to be simplified or removed.

Note that this doesn't yet allow for multi-block directories to be
written.  The next PR will update `folderBranchOps` and
`folderUpdatePrepper` to make that happen.

Issue: KBFS-3302