folder_update_prepper: use `dirData` when prepping paths
This updates `folderUpdatePrepper` to use `dirData` when updating dir
entries, and to ready any dirty blocks within the directory when
prepping the paths for commit.

The ready path() isn't yet exercised in any tests though; a future PR
may end up having to fix bugs here.  It's split out in this way,
pre-testing, for ease of review.

Issue: KBFS-3302