folder_block_ops: improve locking situation with dirData
When fetching child directory blocks in parallel in `dirData` (via
`blockTree`), we need to hold `blockLock` the whole time.  That means
we have to lock `blockLock` when we create the `dirData`, and release
it when we're done with that instance.

So if it's not already locked, have the `dirData` constructor method
return an undo function that must be called when the `dirData` goes
out of scope, which unlocks `blockLock`.

And fix up all other `folderBlockOps` calls that previously locked it
during the course of conflict resolution.

Issue: KBFS-3303