folder_update_prepper: estimate directory size of indirect blocks
Now that dirs can have multiple levels of indirection, it doesn't make
sense to always use the plain (encoded, but not encrypted) size of the
top block as the `DirEntry.Size`.  But it's probably too expensive to
download all leaf blocks, and encode them, to find out the "correct"
value during every folder update.

Instead, use a simple heuristic based on the number of indirect
pointers that isn't great, but probably good enough.

Issue: KBFS-3305