ioutil: quick workaround to try to reduce test flakes
TestJournalCoalescing* tests have a problem where sometimes
`os.MkdirAll()` calls return a "No such file or directory", which
doesn't make much sense.  I suspect it's a race where a directory
exists, but then is wiped before the next child can be made, or
something like that.  I don't know what is deleting those directories
though, whether it's something in Jenkins or something in the code or
test.

For now, let's try re-trying the `MkdirAll` call once, to see if it
helps.  If not, we can drill deeper into exactly what's going on.

Issue: KBFS-3245