libgit: don't hold server locks for GC for too long
Instead, refactor the autogit strategy and only create a "working"
file while holding the lock.  If another worker has already created
the file and is within an hour time limit according to the timestamp
on the file, just return `nil` to avoid showing an error to the user.
If the other worker fails, they'll get another message about doing GC
on a later repo operation.

Issue: KBFS-3407