cr_chains: use default path for unknown, deleted-chain ops
When an entry is both created and deleted within a set of operations
across a set of MD updates, we won't be able to find its corresponding
`rmOp` using the `mostRecent` pointer, because that pointer doesn't
need to be unreferenced.  Instead, just use a default using the root
pointer.

Issue: KBFS-3408
Issue: KBFS-3414 (maybe)