tlf_handle: don't explicitly check team membership while migrating
There's no need to explicitly check team membership while migrating an
existing folder to an implicit team, because the name of the folder is
already validated against the folder's current MD update during
`MDOps.GetIDForHandle()`.

Plus, checking the membership here needlessly fails in some cases,
such as for PUK-less members who appear in the team name, but would
not be reported by the service as resolved team members.

Issue: KBFS-3441