disk_md_cache: introduce disk cache for the last known MD heads
The cache has a Stage+Commit flow, instead of the usual Put, because
according to the KBFS Offline design, we might not want to cache the
MD right after learning about it.  Instead, we might want to wait
until syncing is done, or until the user actually accesses it for the
first time.  In the meantime, the MD must be stored in encrypted form
somewhere, so this cache stages them in memory until they are fully
commit to disk or unstaged.

Issue: KBFS-3503