kbfsgit: resolve sym refs and use dest ref names in ref data map
When we construct the ref-data-by-name map to `libgit.UpdateRepoMD`,
we were erroring out when trying to look up the hash corresponding to
symbolic references like "HEAD".  We should resolve those before we
look up the hash.

Also, we were putting the _source_ reference name as the key to the
map, rather than the dest name.  That meant if someone pushed from a
local branch directly to master, that would not be recorded as a
master push in the server's DB.  Fix that by always using the
destination ref name instead.

Issue: keybase/client#14423