libfuse: on macOS, delete a moved node from the `nodes` cache
Otherwise, the kernel might still treat it as having the old parent
ID.

This shouldn't mess up any kernel caching or libkbfs garbage
collection, since if the original node for the moved node is still in
use, bazil.org would hold a reference to it.

But it's definitely a little weird that the same node might have two
inodes pointing into it.  We should remove this once the osxfuse issue
is fixed.

Issue: #1881
Issue: KBFS-3576
Issue: osxfuse/osxfuse#553