mdserver_remote: get latest handle from cache when disconnected
This RPC is on the critical path for parsing implicit team handles, so
needs to be available offline.  But in the offline case, the latest
cached MD should be good enough.  The call is only there to prevent
using an old TLF ID from a TLF that's already been finalized.

Issue: KBFS-3579