prefetcher: reschedule prefetches with backoff when disk is full
Instead of canceling prefetches when the disk is almost full,
reschedule them (starting with the top-most block) using an
exponential-backoff timer.  This way, the user can see a warning that
the disk is full, make some space, and the prefetches will
automatically resume (eventually).

If needed in the future, we can add a background process that checks
for more disk space and then reschedules them immediately.  But I
worry about racing with other processes needing the disk more
immediately, so let's see how this works for now.

Issue: KBFS-3519