folder_branch_ops: support partial sync config
We want to cache the list of partially-synced paths in an encrypted
block on disk.  Since that has to happen in folderBranchOps, this
moves the main interface for setting/getting sync config into the
KBFSOps interface.  (Of course the paths, including this exact sync
config list, is still logged, but at least logs can be easily deleted
without side effects, while the sync config DB cannot.)

In the lower-level interface, we move from storing the keybase1 sync
config object in the DB (which contains the plaintext path list) to
storing a new struct which contains a block pointer, buffer, and key
half.  This struct is backwards compatible with any current sync
configs stored using the old struct.

We don't yet act on this partial sync config; that will happen in a
separate PR.

Issue: KBFS-3522