prefetcher: fix a few action-related things
1. If a child prefetch will change the action of an existing prefetch
for that child, make sure to send a new request to the main
prefetching worker.

2. Introduce a new bit for explicit deep-syncing; otherwise combining
a sync and a prefetch action will end up kicking off an unintended
deep sync.

3. The prefetcher should request the block before we decide whether to
skip the prefetch, to make sure it gets moved into the correct cache
(important in a future PR).

4. More block request action helper functions for turning an action
solo, and adding sync capabilities to it.

Issue: KBFS-3523