prefetcher: prefetch up to tail blocks when syncing
When prefetching an indirect block and syncing is enabled, we should
prefetch all the way down to the tail blocks, because it doesn't make
sense to sync just part of a multi-block object.

This will be important for partially-synced paths.  For example, if
syncing path `a/b/c`, and directory `a` is a multi-block directory, we
should sync all of the blocks of `a`.  That way `ls a` can succeed
while offline, while the user is navigating up to `a/b/c`.