simplefs: use `NewFSIfExists` for read-only operations
We don't want to accidentally create a TLF when reading an empty TLF,
so use `libfs.NewFSIfExists` in that case.

It would be nice to also use a read-only FS too, but if a
`libkfbs.Node` gets created in read-only mode, it will be served to
subsequent calls as read-only as well, even if they're done through a
read-write file system.  Maybe that's something we want to fix in the
future, but for now just make sure the TLF isn't created.